requirements.txt
# Geodesyxx Reproduction Package Dependencies
# Exact versions used for ICLR 2024 paper results

# Core Dependencies
torch==2.1.0
numpy==1.24.3
scipy==1.11.4
pandas==2.0.3

# Transformer Models and NLP
transformers==4.35.2
datasets==2.14.6
tokenizers==0.15.0
huggingface-hub==0.17.3
safetensors==0.4.0

# Machine Learning Utilities  
scikit-learn==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.66.1

# Statistical Analysis
statsmodels==0.14.0
pingouin==0.5.3

# Configuration Management
pyyaml==6.0.1
omegaconf==2.3.0

# Data Processing
nltk==3.8.1  # For WordNet experiments

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0

# Memory Profiling (Optional)
psutil==5.9.6
memory-profiler==0.61.0

# Platform-Specific Notes:
# - Apple Silicon: Default PyTorch installation includes MPS support
# - NVIDIA GPU: Use torch with CUDA 11.8 support  
# - CPU only: Default installation sufficient

# Installation Instructions:
# 1. Create virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    venv\Scripts\activate     # Windows
#
# 2. Install PyTorch with appropriate backend:
#    # Apple Silicon (recommended)
#    pip install torch torchvision torchaudio
#    
#    # NVIDIA GPU (CUDA 11.8)
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#    
#    # CPU only
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# 3. Install remaining requirements:
#    pip install -r requirements.txt
#
# 4. Download required datasets:
#    python data/download_datasets.py

# Version Compatibility Notes:
# - Python 3.8+ required (tested on 3.10.12)
# - PyTorch 2.0+ required for proper MPS support
# - Transformers 4.30+ required for DistilBERT compatibility
# - NumPy <2.0 for compatibility with older packages